make read_tensor_data free function
