+++ src/core/include/openvino/runtime/tensor.hpp
@@ -254,20 +254,6 @@ public:
         T::type_check(*this);
         return *static_cast<const T*>(this);
     }
-
-    /// \brief Read a tensor content from a file. Only raw data is loaded.
-    /// \param file_name Path to file to read.
-    /// \param element_type Element type, when not specified the it is assumed as element::u8.
-    /// \param shape Shape for resulting tensor. If provided shape is static, specified number of elements is read only.
-    ///              File should contain enough bytes, an exception is raised otherwise.
-    ///              One of the dimensions can be dynamic. In this case it will be determined automatically based on the
-    ///              length of the file content and `offset`. Default value is [?].
-    /// \param offset_in_bytes Read file starting from specified offset. Default is 0. The remining size of the file
-    /// should be compatible with shape.
-    static Tensor from_file(const std::filesystem::path& file_name,
-                            const element::Type& element_type = element::u8,
-                            const PartialShape& shape = PartialShape::dynamic(1),
-                            std::size_t offset_in_bytes = 0);
 };
 
 /**
@@ -275,4 +261,18 @@ public:
  */
 using TensorVector = std::vector<Tensor>;
 
+/// \brief Read a tensor content from a file. Only raw data is loaded.
+/// \param file_name Path to file to read.
+/// \param element_type Element type, when not specified the it is assumed as element::u8.
+/// \param shape Shape for resulting tensor. If provided shape is static, specified number of elements is read only.
+///              File should contain enough bytes, an exception is raised otherwise.
+///              One of the dimensions can be dynamic. In this case it will be determined automatically based on the
+///              length of the file content and `offset`. Default value is [?].
+/// \param offset_in_bytes Read file starting from specified offset. Default is 0. The remining size of the file should
+/// be compatible with shape.
+OPENVINO_API
+Tensor read_tensor_data(const std::filesystem::path& file_name,
+                        const element::Type& element_type = element::u8,
+                        const PartialShape& shape = PartialShape::dynamic(1),
+                        std::size_t offset_in_bytes = 0);
 }  // namespace ov
